Downgrade SwiftLint's `todo` rule to `warning` severity
It should be possible to add `TODO` and `FIXME` comments during development and still be able to build the app.

That said, the CI should be configured to promote the `todo` rule to `error` severity on the `master` branch and in PRs against to to prevent such comments from entering production.